Add dedicated Windows test
