+++ pkg/trace/api/api_nix_test.go
@@ -17,8 +17,10 @@ import (
 	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
+	"github.com/DataDog/datadog-agent/pkg/trace/log"
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
 	"github.com/DataDog/datadog-agent/pkg/trace/testutil"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestUDS(t *testing.T) {
@@ -70,4 +72,54 @@ func TestUDS(t *testing.T) {
 			t.Fatalf("expected http.StatusOK, got response: %#v", resp)
 		}
 	})
+}
+
+func TestHTTPReceiverStart(t *testing.T) {
+	var logs bytes.Buffer
+	old := log.SetLogger(log.NewBufferLogger(&logs))
+	defer log.SetLogger(old)
+
+	for name, tt := range map[string]struct {
+		port   int      // receiver port
+		socket string   // socket
+		out    []string // expected log output (uses strings.Contains)
+	}{
+		"off": {
+			out: []string{"HTTP Server is off: all listeners are disabled"},
+		},
+		"tcp": {
+			port: 8129,
+			out: []string{
+				"Listening for traces at http://localhost:8129",
+			},
+		},
+		"uds": {
+			socket: "/tmp/agent.sock",
+			out: []string{
+				"HTTP receiver disabled by config (apm_config.receiver_port: 0)",
+				"Listening for traces at unix:///tmp/agent.sock",
+			},
+		},
+		"both": {
+			port:   8129,
+			socket: "/tmp/agent.sock",
+			out: []string{
+				"Listening for traces at http://localhost:8129",
+				"Listening for traces at unix:///tmp/agent.sock",
+			},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			logs.Reset()
+			cfg := config.New()
+			cfg.ReceiverPort = tt.port
+			cfg.ReceiverSocket = tt.socket
+			r := newTestReceiverFromConfig(cfg)
+			r.Start()
+			defer r.Stop()
+			for _, l := range tt.out {
+				assert.Contains(t, logs.String(), l)
+			}
+		})
+	}
 }
