+++ pkg/trace/api/pipe.go
@@ -15,8 +15,8 @@ import (
 )
 
 // listenPipe returns a listener on the given Windows Pipe, using the provided security
-// descriptor and buffer size; replaced in tests.
-var listenPipe = func(path string, secdec string, bufferSize int) (net.Listener, error) {
+// descriptor and buffer size.
+func listenPipe(path string, secdec string, bufferSize int) (net.Listener, error) {
 	ln, err := winio.ListenPipe(path, &winio.PipeConfig{
 		SecurityDescriptor: secdec,
 		InputBufferSize:    int32(bufferSize),
