+++ pkg/trace/api/pipe_off.go
@@ -13,7 +13,7 @@ import (
 	"net"
 )
 
-// listenPipe return a nil-listener and an error on non-Windows operating systems; replaced in tests.
-var listenPipe = func(_, _ string, _ int) (net.Listener, error) {
+// listenPipe return a nil-listener and an error on non-Windows operating systems.
+func listenPipe(_, _ string, _ int) (net.Listener, error) {
 	return nil, errors.New("Windows named pipes are only supported on Windows operating systems")
 }
